requirements.txt
pandas
google-cloud-bigquery
numpy
requests
openpyxl
matplotlib
seaborn